implementacion de clase PersonaLogic
