+++ ejemplos/Ej5AccesoADatos/logic/Login.java
@@ -24,24 +24,5 @@ public class Login {
 		return dp.getAll();
 	}
 
-	public Persona getByDocumento(Persona per) {
-		return dp.getByDocumento(per);
-		
-	}
-	
-	public LinkedList<Persona> getByApellido(Persona per){
-		return dp.getByApellido(per);
-	}
-	
-	public void addPersona(Persona p) {
-		dp.add(p);
-	}
-	
-	public Persona editPersona(Persona pOld, Persona p) {
-		return dp.editPersona(pOld, p);
-	}
-	
-	public void eliminarPersona(Persona p) {
-		dp.eliminarPersona(p);
-	}
+
 }
